WithSearch Tweaks (#2163)
* Migrating Search Methods into Separate Traits

* Fix styling

* PHPStan Fixes

* Fix styling

---------
