+++ src/Traits/Core/Search/HandlesSearchFieldStyling.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Rappasoft\LaravelLivewireTables\Traits\Core\Search;
+
+use Rappasoft\LaravelLivewireTables\Traits\Core\Search\Styling\{HasSearchIcon, HasSearchInput};
+
+trait HandlesSearchFieldStyling
+{
+    use HasSearchIcon,
+        HasSearchInput;
+}
